CPPRunner is able to run BA modules
